Modified Files
Did some tech debt:

Created a singleton HTTP client for the base data access

Modified database so that it stores the salt, hash and work factor for a
password.

Added an encryption helper class for password validation and generation.

Removed some empty projects that were sitting around.

Renamed a bunch of methods and objects to suit their context better.